Fix three bugs
This patch fixes 3 bugs that were present but were undiscovered when
setting up a smaller experiment. The first one is the use of the
variable 'active' when it was not yet assigned (when an experiment was
not swapped in yet). The other is when a node has multiple
interfaces. The parsing did not happen as it should. The last bug
happened because of Python's weak typing, causing the same 'interface'
object being copied to all links. This resulted in weird behaviour as
all links got the same IP and OS specific name.